requirements.txt
click==6.7
Flask==0.12.2
gpio==0.2.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pi-sht1x==1.0.9
python-dateutil==2.6.1
redis==2.10.6
RPi.GPIO==0.6.3
six==1.11.0
uWSGI==2.0.15
Werkzeug==0.13
